% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/halcyon_app.R
\name{halcyon_app}
\alias{halcyon_app}
\title{Halcyon App}
\usage{
halcyon_app(outfmt10, ocf = function(x) {     as.numeric(quantile(x)[4]) * 1.5
  })
}
\arguments{
\item{outfmt10}{A data frame imported with \code{read_outfmt10}.}

\item{ocf}{Outlier calculation function}
}
\description{
Halcyon App
}

