% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/m_score.R
\name{m_score}
\alias{m_score}
\title{Calculate the m-score for each element in a numeric vector}
\usage{
m_score(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
The m-score is calculated by the following formula:
For each element i in vector x calculate (i - median(x))/mad(x). This
function returns a vector of m-scores corresponding to the input vector.
}
\examples{
\dontrun{

m_score(1:10)
m_score(rnorm(100))

}
}

