% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_outfmt10.R
\name{read_outfmt10}
\alias{read_outfmt10}
\title{Import outfmt 10 result}
\usage{
read_outfmt10(path = ifelse(interactive(), file.choose(), NULL))
}
\arguments{
\item{path}{The path to the blast result.}
}
\description{
Convert a result from using \code{blastn} with the \code{-outfmt "10"} flag
to a data frame.
}
\examples{
\dontrun{

# Choose the path to the file interactively
read_outfmt10()

# Or explicitly specify a path
read_outfmt10(file.path("path", "to", "file.out"))

}
}

